backend/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.1
torch>=2.1
shap>=0.44
fastapi>=0.104
uvicorn>=0.24
pydantic>=2.4
python-jose[cryptography]>=3.3
langgraph>=0.1
langchain>=0.1
langchain-anthropic>=0.1
langchain-openai>=0.1
httpx>=0.25
pytest>=7.4
pyyaml>=6.0
python-dotenv>=1.0
joblib>=1.3
matplotlib>=3.7
plotly>=5.17
